Separate bridge-ralated implementation in ExecDelivery
